Add an installer
This change adds a setup.py file that lets users install the library
with pip, rather than needing to modify their PYTHONPATH environment
variable.

To support this, the files in lib/ get moved to pyvirtualbench/,
.gitignore is modified to ignore the created installation, and
the README.md now has instructions on how to use pip to install
the library for both users and developers.